ReadSingletoDB functions
Currently, one script exists that processes the words and uses from one file and loads them into the table in the database. It is built assuming that the words in the database are not already there, and a script which updates the old table must account for this going forwards.